refactor txpool options (#5512)
* refactor txpool options

* move options in BesuCommand

---------
